LPD-36199 Move to correct place
