+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/miller_columns/constants/acceptingTypes.ts
@@ -6,8 +6,3 @@
 export const ACCEPTING_TYPES = {
 	ITEM: 'ITEM',
 };
-
-// Defines the distance in px from the border to the center of an item where
-// dropZone changes from TOP/BOTTOM to ELEMENT
-
-export const ITEM_HOVER_BORDER_LIMIT = 20;
